+++ Libplanet/Blockchain/Policies/IBlockPolicy.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using Libplanet.Action;
 using Libplanet.Blocks;
@@ -12,28 +11,10 @@ namespace Libplanet.Blockchain.Policies
     /// </summary>
     /// <typeparam name="T">An <see cref="IAction"/> type.  It should match
     /// to <see cref="Block{T}"/>'s type parameter.</typeparam>
+    /// <seealso cref="IBlockPolicyExtension"/>
     public interface IBlockPolicy<T>
         where T : IAction, new()
     {
-        /// <summary>
-        /// Checks if <paramref name="blocks"/> are invalid, and if that
-        /// returns the reason.
-        /// <para>Note that it returns <c>null</c> when blocks are
-        /// <em>valid</em>.</para>
-        /// </summary>
-        /// <param name="blocks">Consecutive <see cref="Block{T}"/>s to
-        /// validate.</param>
-        /// <param name="currentTime">The current time to be used to validate
-        /// of <see cref="Block{T}.Timestamp"/>s.
-        /// Usually <see cref="DateTimeOffset.UtcNow"/> is used.</param>
-        /// <returns>The reason why the given <paramref name="blocks"/> are
-        /// <em>invalid</em>, or <c>null</c> if <paramref name="blocks"/> are
-        /// <em>valid</em>.</returns>
-        InvalidBlockException ValidateBlocks(
-            IReadOnlyList<Block<T>> blocks,
-            DateTimeOffset currentTime
-        );
-
         /// <summary>
         /// Checks if <paramref name="nextBlock"/> is invalid, and if that
         /// returns the reason.
@@ -47,6 +28,7 @@ namespace Libplanet.Blockchain.Policies
         /// <returns>The reason why the given <paramref name="blocks"/> are
         /// <em>invalid</em>, or <c>null</c> if <paramref name="blocks"/> are
         /// <em>valid</em>.</returns>
+        /// <seealso cref="IBlockPolicyExtension.ValidateBlocks{T}"/>
         InvalidBlockException ValidateNextBlock(
             IReadOnlyList<Block<T>> blocks,
             Block<T> nextBlock);